requirements.txt
streamlit
qdrant-client
langchain
langchain-community
langchain-google-genai
sentence-transformers
python-dotenv
